+++ src/constants.js
@@ -1,0 +1,44 @@
+/* constants.js */
+
+// Element IDs
+
+export const SKIP_TO_ID            = 'id-skip-to-ver-5';
+export const SKIP_TO_MENU_STYLE_ID = 'id-skip-to-menu-style';
+
+export const SCRIPT_EXTENSION_ID   = `id-skip-to-extension`;
+export const SCRIPT_BOOKMARKLET_ID = `id-skip-to-bookmarklet`;
+
+export const MENU_ID   = 'id-skip-to-menu';
+
+export const MENU_LANDMARK_GROUP_ID        = 'id-skip-to-landmark-group';
+export const MENU_LANDMARK_GROUP_LABEL_ID  = 'id-skip-to-landmark-group-label';
+
+export const MENU_HEADINGS_GROUP_ID        = 'id-skip-to-heading-group';
+export const MENU_HEADINGS_GROUP_LABEL_ID  = 'id-skip-to-heading-group-label';
+
+export const MENU_SHORTCUTS_GROUP_ID       = 'id-skip-to-shortcuts-group';
+export const MENU_SHORTCUTS_GROUP_LABEL_ID = 'id-skip-to-shortcuts-group-label';
+
+export const MESSAGE_ID = 'id-skip-to-message';
+
+export const HIGHLIGHT_ID = 'id-skip-to-highlight-overlay';
+
+// Custom element names
+
+export const PAGE_SCRIPT_ELEMENT_NAME = 'skip-to-content';
+export const BOOKMARKLET_ELEMENT_NAME = 'skip-to-content-bookmarklet';
+export const EXTENSION_ELEMENT_NAME   = 'skip-to-content-extension';
+
+export const INFO_DIALOG_ELEMENT_NAME = 'skip-to-content-info-dialog-574';
+export const MESSAGE_ELEMENT_NAME     = 'skip-to-content-message-element-574';
+export const HIGHLIGHT_ELEMENT_NAME   = 'skip-to-content-highlight-element-574';
+
+// Attributes
+
+export const ATTR_SKIP_TO_DATA = 'data-skipto';
+
+// URLs to more information
+
+export const MORE_PAGE_SCRIPT_INFO_URL ='https://skipto-landmarks-headings.github.io/page-script-5/';
+export const MORE_SHORTCUT_INFO_URL    ='https://skipto-landmarks-headings.github.io/page-script-5/shortcuts.html';
+