+++ src/skiptoContent.js
@@ -6,6 +6,13 @@ import SkiptoMenuButton from './skiptoMenuButton.js';
 import DebugLogging  from './debug.js';
 
 import {
+  ATTR_SKIP_TO_DATA,
+  INFO_DIALOG_ELEMENT_NAME,
+  MESSAGE_ELEMENT_NAME,
+  HIGHLIGHT_ELEMENT_NAME
+} from './constants.js';
+
+import {
   getLandmarksAndHeadings
 } from './landmarksHeadings.js';
 
@@ -17,15 +24,17 @@ import {
 const debug = new DebugLogging('skiptoContent', false);
 debug.flag = false;
 
+/* @class SkipToContent574
+ *
+ */
 
-export default class SkipToContent573 extends HTMLElement {
+export default class SkipToContent574 extends HTMLElement {
 
   constructor() {
     // Always call super first in constructor
     super();
     this.attachShadow({ mode: 'open' });
-    this.skipToId = 'id-skip-to';
-    this.version = "5.7.3";
+    this.version = "5.7.4";
     this.buttonSkipTo = false;
     this.initialized = false;
 
@@ -182,7 +191,7 @@ export default class SkipToContent573 extends HTMLElement {
 
   static get observedAttributes() {
     return [
-      "data-skipto",
+      ATTR_SKIP_TO_DATA,
       "setfocus",
       "type",
       "shortcuts",
@@ -192,7 +201,7 @@ export default class SkipToContent573 extends HTMLElement {
 
   attributeChangedCallback(name, oldValue, newValue) {
 
-    if (name === 'data-skipto') {
+    if (name === ATTR_SKIP_TO_DATA) {
       this.config = this.setupConfigFromDataAttribute(this.config, newValue);
     }
 
@@ -257,9 +266,9 @@ export default class SkipToContent573 extends HTMLElement {
       }
 
       // Check for data-skipto attribute values for configuration
-      const configElem = document.querySelector('[data-skipto]');
+      const configElem = document.querySelector(`[${ATTR_SKIP_TO_DATA}]`);
       if (configElem) {
-        const params = configElem.getAttribute('data-skipto');
+        const params = configElem.getAttribute(ATTR_SKIP_TO_DATA);
         this.config  = this.setupConfigFromDataAttribute(this.config, params);
       }
 
@@ -364,19 +373,28 @@ export default class SkipToContent573 extends HTMLElement {
     }
 
     renderStyleElement(this.shadowRoot, config, this.skipToId);
+
     if (this.buttonSkipTo) {
       this.buttonSkipTo.updateLabels(config);
       this.buttonSkipTo.setDisplayOption(config['displayOption']);
     }
 
-    const infoDialog = document.querySelector('skip-to-shortcuts-info-dialog');
+    const infoDialog = document.querySelector(INFO_DIALOG_ELEMENT_NAME);
+    debug.flag && debug.log(`[infoDialog]: ${infoDialog}`);
     if (infoDialog) {
       infoDialog.configureStyle(config);
     }
 
-    const shortcutsMessage = document.querySelector('skip-to-shortcuts-message');
+    const shortcutsMessage = document.querySelector(MESSAGE_ELEMENT_NAME);
+    debug.flag && debug.log(`[shortcutMessage]: ${shortcutsMessage}`);
     if (shortcutsMessage) {
       shortcutsMessage.configureStyle(config);
+    }
+
+    const highlight = document.querySelector(HIGHLIGHT_ELEMENT_NAME);
+    debug.flag && debug.log(`[highlight]: ${highlight}`);
+    if (highlight) {
+      highlight.configureStyle(config);
     }
 
     return config;